requirements.txt
flask
openai>=1.54.0
pymysql
pdfplumber
gunicorn
python-dotenv
numpy
Flask-Limiter
Pillow
psycopg2-binary